idle mmc
git-svn-id: https://cc3.svn.sourceforge.net/svnroot/cc3/trunk@372 111db651-790d-0410-a044-b0abe44a5820

+++ hal/lpc2106-cmucam3/rdcf2.h
@@ -106,8 +106,8 @@ struct rdcf
   /* values that must be initialized by the calling program */
   union  IO_BUFFER buffer;
   // hardware access.
-  bool (*ReadSector)(long sector, uint8_t * buf);
-  bool (*WriteSector)(long sector, const uint8_t * buf);
+  bool (*ReadSector)(uint32_t sector, uint8_t * buf);
+  bool (*WriteSector)(uint32_t sector, const uint8_t * buf);
   /* file information */
   struct rdcf_file_information file;
   /* result codes */
